move modules from registry/ to utils/
